Extract "common_pipes" to "tests/common_pipes.py"
* Similar to the description in the README.
